+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
-import Admin from "./components/Admin";
-import Home from "./components/Home";
+import Admin from "./components/screens/Admin";
+import Home from "./components/screens/Home";
 
 export default class App extends Component {
   render() {
