+++ frontend/src/components/screencomponents/homecomponent/Testimonial.js
@@ -1,0 +1,120 @@
+import React from "react";
+
+function Testimonial() {
+  return (
+    <>
+      <section id="testimonials" className="section-bg">
+          <div className="container" data-aso="zoom-in">
+            <header className="section-header">
+              <h3>Testimonials</h3>
+            </header>
+
+            <div className="row justify-content-center">
+              <div className="col-lg-8">
+                <div
+                  className="testimonials-slider swiper-container"
+                  data-aos="fade-up"
+                  data-aos-delay="100"
+                >
+                  <div className="swiper-wrapper">
+                    <div className="swiper-slide">
+                      <div className="testimonial-item">
+                        <img
+                          src="assets/img/testimonial-1.jpg"
+                          className="testimonial-img"
+                          alt=""
+                        />
+                        <h3>Saul Goodman</h3>
+                        <h4>Ceo &amp; Founder</h4>
+                        <p>
+                          Proin iaculis purus consequat sem cure digni ssim
+                          donec porttitora entum suscipit rhoncus. Accusantium
+                          quam, ultricies eget id, aliquam eget nibh et. Maecen
+                          aliquam, risus at semper.
+                        </p>
+                      </div>
+                    </div>
+
+                    <div className="swiper-slide">
+                      <div className="testimonial-item">
+                        <img
+                          src="assets/img/testimonial-2.jpg"
+                          className="testimonial-img"
+                          alt=""
+                        />
+                        <h3>Sara Wilsson</h3>
+                        <h4>Designer</h4>
+                        <p>
+                          Export tempor illum tamen malis malis eram quae irure
+                          esse labore quem cillum quid cillum eram malis quorum
+                          velit fore eram velit sunt aliqua noster fugiat irure
+                          amet legam anim culpa.
+                        </p>
+                      </div>
+                    </div>
+
+                    <div className="swiper-slide">
+                      <div className="testimonial-item">
+                        <img
+                          src="assets/img/testimonial-3.jpg"
+                          className="testimonial-img"
+                          alt=""
+                        />
+                        <h3>Jena Karlis</h3>
+                        <h4>Store Owner</h4>
+                        <p>
+                          Enim nisi quem export duis labore cillum quae magna
+                          enim sint quorum nulla quem veniam duis minim tempor
+                          labore quem eram duis noster aute amet eram fore quis
+                          sint minim.
+                        </p>
+                      </div>
+                    </div>
+
+                    <div className="swiper-slide">
+                      <div className="testimonial-item">
+                        <img
+                          src="assets/img/testimonial-4.jpg"
+                          className="testimonial-img"
+                          alt=""
+                        />
+                        <h3>Matt Brandon</h3>
+                        <h4>Freelancer</h4>
+                        <p>
+                          Fugiat enim eram quae cillum dolore dolor amet nulla
+                          culpa multos export minim fugiat minim velit minim
+                          dolor enim duis veniam ipsum anim magna sunt elit fore
+                          quem dolore labore illum veniam.
+                        </p>
+                      </div>
+                    </div>
+                    <div className="swiper-slide">
+                      <div className="testimonial-item">
+                        <img
+                          src="assets/img/testimonial-5.jpg"
+                          className="testimonial-img"
+                          alt=""
+                        />
+                        <h3>John Larson</h3>
+                        <h4>Entrepreneur</h4>
+                        <p>
+                          Quis quorum aliqua sint quem legam fore sunt eram
+                          irure aliqua veniam tempor noster veniam enim culpa
+                          labore duis sunt culpa nulla illum cillum fugiat legam
+                          esse veniam culpa fore nisi cillum quid.
+                        </p>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="swiper-pagination"></div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </section>
+
+    </>
+  );
+}
+
+export default Testimonial;
